SUBSWARM:  Added subswarm assignment to the dashboard
Added a button and spinBox to the dashboard to assign vehicles to subswarms.
The functionality of the table was changed to allow the selection of multiple
rows to allow the assignment of multiple aircraft to a subswarm with a single
button push.  The behavior of the Start Following button was modified some
to prevent sending a message to an aircraft when nothing is selected.  The
change also allows selection of multiple aircraft to which a start following
message will be sent.