qt_gui:  Subswarm assignment functionality
Added functionality to the dashboard to assign aircraft to subswarms.  The
selection functionality of the table was modified to allow selection of
multiple rows (UAVs) at once so that they can all be assigned to a single
subswarm with one button push.  Also modified the callback for the
"Start Following" button to use this functionality.