state:  FlightStatus message format changes
The gps_sats field was removed from the FlightStatus message.  The
gps_ok field is now parsed instead when FlightStatus messages are
received.

Note:  Also requires an update to qt_gui to prevent an error from bein
generated by the dashboard (see next commit).

Also: Parsed the (newly implemented) swarm_state value from the FlightStatus
message so that it can be displayed on the dashboard.