qt_gui:  Dashboard column modifications
Switched the GPS column on the dashboard to display red/green based on
the value of the gps_ok field in the FlightStatus message (gps_sats
was removed from the message).

Added swarm state to the dashboard display.  Per the mission flow
discussions that have been ongoing, swarm state describes what portion
of the mission the vehicle is in.  A vehicle can only be assigned to a
subswarm if it is in the IN_SWARM state.

Note:  there's some hacky stuff in one file that's just there until I
figure out a cleaner way to get the import to work without a crash.