ACS_NETWORK:  Added new SwarmState message to the network module
Implemented a new SwarmState message to the network module for use in
manually setting a UAV swarm state.  This will not usually be required as the
vehicle will automatically detect transitions from one state to another, but
during testing it may be desirable to manually set the swarm state (mostly
in cases where we want to go from "egress" or "landing" back to "swarm ready".