qt_gui: Manual swarm state reset, control/swarm mode display, & layout reorg
Added the ability to manually reset the swarm state of selected vehicles. This
will not normally be required, but may prove useful if we want to resume
swarm testing after sending the aircraft to the "egress" state.
Added columns to the dashboard display for active controller and swarm
behavior.  These should make tracking what is going on with individual
aircraft easier.
Rearranged the dashboard some so that swarm-related functionality is grouped
in one place and manual override (which probably won't be used in "normal"
situations) is grouped in another.