qt_gui: Color warning for autopilot mode chng. Also fixed batt_rem bug.
Added color coding to the "mode" field in the dashboard table to make
unexpected modes more apparent.  No color (good) indicates that the autopilot
is in "auto" mode.  Most other modes will cause the mode field to turn
yellow.  If the mode switches to "manual", the field will turn red.  An
aural warning was also implemented for changes to "manual", but it only
worked sporadically, so it is commented out for now.

Also fixed the bug noticed in the field that resulted in battery remaining
being reported as "0" (it was a misnamed variable).  Switched battery
remaining field value to voltage vice percent.  It's an easy switch to change
it back, but voltage seems to be more useful than percent, as that is what
the failsafe and "bingo" are based on.