qt_gui:  Swarm behavior msgs for ordering different behaviors w/ parameters
Added functionality to order different swarm behaviors that require
behavior-specific parameters.  Functionality consists of popup dialogs for
entering parameters for the swarm sequence landing behavior (landing
direction and runway--e.g, "RW 28 (East to West)") and the fixed formation
behavior (distance from leader, angle off the leader's bow of the ordered
position, and a Boolean for whether or not all followers are to take the
same position relative to the single leader).  A menu item was added to the
dashboard to select which behavior type to initiate.

For the swarm sequence landing behavior, the underlying message specifies
the ID of the first waypoint in the landing sequence (meaning the ID is
hard coded in the SwarmCommander to align with the selected landing option).
The swarm manager on the vehicle will verify that this is the correct type
of waypoint (189) before successfully activating the behavior.