qt_gui: Stale & on-ground UAVs filtered from dashboard
UAVs that are not airborne (preflight, flight ready, or on deck swarm state)
will no longer be displayed on the dashboard.  Also, aircraft that do not
provide update messages for more than 1 minute will be filtered.

GPS status and link status columns are now hidden.

Both of these changes are intended to reduce clutter and unnecessary
information to the swarm operator as we scale up numbers.

Also changed verbiage on the landing sequencer pulldown to "Land A" and
"Land B".